requirements.txt
apache-airflow==2.7.0
apache-airflow-providers-common-sql==1.7.1
apache-airflow-providers-http==4.5.1
apache-airflow-providers-ftp==3.5.1
kafka-python==2.1.1
requests==2.31.0
psycopg2-binary==2.9.6
websocket-client
pyspark==3.4.0
cassandra-driver==3.27.0
cassandra-driver==3.27.0

